+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2022-03-27
+
+### 🔥 Added
+
+- `project` repo CRUD
+- `project` controller & router REST API CRUD
+- `partnumber` repo CRUD
+- `partnumber` controller & router REST API CRUD
+
+## [2.1.0] - 2022-03-24
+
+### 🔥 Added
+
+- `industry` repo CRUD
+- `industry` controller & router REST API CRUD
+
+## [2.0.0] - 2022-03-23
+
+### 🔥 Added
+
+- `project-client` repo CRUD
+- `projectClient` controller & router REST API CRUD
+
 ## [2.0.0] - 2022-03-18
 
 ### 🔥 Added